Fix for Kitkat photo save

+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/sensor/cbt/CompartmentBagMainActivity.java
@@ -1,24 +1,32 @@
 package org.akvo.caddisfly.sensor.cbt;
 
+import android.Manifest;
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.Dialog;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.graphics.Color;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.provider.MediaStore;
 import android.support.annotation.NonNull;
+import android.support.design.widget.CoordinatorLayout;
+import android.support.design.widget.Snackbar;
+import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.DialogFragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v4.content.FileProvider;
 import android.support.v7.app.AlertDialog;
 import android.util.SparseArray;
+import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import org.akvo.caddisfly.R;
@@ -29,6 +37,7 @@ import org.akvo.caddisfly.model.MpnValue;
 import org.akvo.caddisfly.model.TestInfo;
 import org.akvo.caddisfly.sensor.SensorConstants;
 import org.akvo.caddisfly.ui.BaseActivity;
+import org.akvo.caddisfly.util.ApiUtil;
 import org.akvo.caddisfly.util.ImageUtil;
 import org.akvo.caddisfly.util.StringUtil;
 import org.json.JSONObject;
@@ -37,12 +46,22 @@ import java.io.File;
 import java.io.IOException;
 import java.util.UUID;
 
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
+import static android.content.pm.PackageManager.PERMISSION_GRANTED;
+
 public class CompartmentBagMainActivity extends BaseActivity implements CompartmentBagFragment.OnFragmentInteractionListener {
 
     public static final int REQUEST_TAKE_PHOTO = 1;
+    private static final int PERMISSION_ALL = 1;
+    private static final String[] PERMISSIONS = {Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE};
+    private static final float SNACK_BAR_LINE_SPACING = 1.4f;
     private final SparseArray<String> results = new SparseArray<>();
     String mCurrentPhotoPath;
     String imageFileName = "";
+    @BindView(R.id.coordinatorLayout)
+    CoordinatorLayout coordinatorLayout;
     private FragmentManager fragmentManager;
     private String mResult = "00000";
 
@@ -50,6 +69,8 @@ public class CompartmentBagMainActivity extends BaseActivity implements Compartm
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_compartment_bag_main);
+
+        ButterKnife.bind(this);
 
         fragmentManager = getSupportFragmentManager();
 
@@ -60,6 +81,14 @@ public class CompartmentBagMainActivity extends BaseActivity implements Compartm
 
     public void onClickNextButton(View view) {
 
+        if (!ApiUtil.hasPermissions(this, PERMISSIONS)) {
+            ActivityCompat.requestPermissions(this, PERMISSIONS, PERMISSION_ALL);
+        } else {
+            startCamera();
+        }
+    }
+
+    public void startCamera() {
         Toast.makeText(this, "Take a photo of the compartment bag", Toast.LENGTH_LONG).show();
 
         Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
@@ -72,17 +101,67 @@ public class CompartmentBagMainActivity extends BaseActivity implements Compartm
             } catch (IOException ex) {
                 // Error occurred while creating the File
             }
+
             // Continue only if the File was successfully created
             if (photoFile != null) {
-                Uri photoURI = FileProvider.getUriForFile(this,
-                        "org.akvo.caddisfly.fileprovider",
-                        photoFile);
+
+                Uri photoURI;
+                if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.KITKAT) {
+                    photoURI = Uri.fromFile(photoFile);
+                } else {
+                    photoURI = FileProvider.getUriForFile(this,
+                            "org.akvo.caddisfly.fileprovider",
+                            photoFile);
+                }
                 takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI);
-                takePictureIntent.putExtra(MediaStore.EXTRA_SIZE_LIMIT, 2000);
                 startActivityForResult(takePictureIntent, REQUEST_TAKE_PHOTO);
             }
         }
     }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
+                                           @NonNull int[] grantResults) {
+
+        final Activity activity = this;
+        if (requestCode == PERMISSION_ALL) {
+            // If request is cancelled, the result arrays are empty.
+            boolean granted = false;
+            for (int grantResult : grantResults) {
+                if (grantResult != PERMISSION_GRANTED) {
+                    granted = false;
+                    break;
+                } else {
+                    granted = true;
+                }
+            }
+            if (granted) {
+                startCamera();
+            } else {
+                Snackbar snackbar = Snackbar
+                        .make(coordinatorLayout, getString(R.string.cameraAndStoragePermissions),
+                                Snackbar.LENGTH_LONG)
+                        .setAction("SETTINGS", new View.OnClickListener() {
+                            @Override
+                            public void onClick(View view) {
+                                ApiUtil.startInstalledAppDetailsActivity(activity);
+                            }
+                        });
+
+                TypedValue typedValue = new TypedValue();
+                getTheme().resolveAttribute(R.attr.colorPrimaryDark, typedValue, true);
+
+                snackbar.setActionTextColor(typedValue.data);
+                View snackView = snackbar.getView();
+                TextView textView = snackView.findViewById(android.support.design.R.id.snackbar_text);
+                textView.setHeight(getResources().getDimensionPixelSize(R.dimen.snackBarHeight));
+                textView.setLineSpacing(0, SNACK_BAR_LINE_SPACING);
+                textView.setTextColor(Color.WHITE);
+                snackbar.show();
+            }
+        }
+    }
+
 
     public void onClickMatchedButton(View view) {
 
@@ -145,15 +224,17 @@ public class CompartmentBagMainActivity extends BaseActivity implements Compartm
 
         ImageUtil.resizeImage(mCurrentPhotoPath, resultImagePath);
 
-        //noinspection ResultOfMethodCallIgnored
-        new File(mCurrentPhotoPath).delete();
+        File imageFile = new File(mCurrentPhotoPath);
+        if (imageFile.exists()) {
+            //noinspection ResultOfMethodCallIgnored
+            new File(mCurrentPhotoPath).delete();
+        }
 
         MpnValue mpnValue = TestConfigHelper.getMpnValueForKey(mResult);
 
         results.put(1, StringUtil.getStringResourceByName(this, mpnValue.getRiskCategory()).toString());
         results.put(2, mpnValue.getMpn());
         results.put(3, mpnValue.getConfidence());
-
 
         JSONObject resultJson = TestConfigHelper.getJsonResult(testInfo, results, -1, imageFileName, null);
         resultIntent.putExtra(SensorConstants.RESPONSE, resultJson.toString());
